Move suspense utils to their own file
